remove sse client and middleware, move to async thunk
