+++ src/modules/oh-indicators-list/store/index.ts
@@ -1,5 +1,6 @@
 import reducer from './slice';
 
 export { selectItems, selectErrorMessage } from './selectors';
-export { getAllItems, indicatorsSlice } from './slice';
+export { getAllItems } from './actions';
+export { indicatorsSlice } from './slice';
 export { reducer as default };
